Fixed the Quantity Change Feature
Modified the remove quantity feature to be able to change the quantity amount of a product between 1-99. There was an issue after modifying it where adding the same product to the cart multiple times would add 2 per click (ex. 1, 3, 5, 7, 9, etc.), so I modified the addToCart function to deal with this issue.